backend/requirements.txt
cs50==9.4.0
Flask==3.1.2
PyJWT==2.10.1
Werkzeug==3.1.3
python-dotenv
Faker
pyfcm==1.5.4
